React API refactoring
- Add a generic useConnectData() hook, to transform an async call or a
  subscription into a result + status tuple.
- Split index.ts into separate files.
- createAppHook() is now a wrapper around useConnectData() that only
  loads the app.
- Improve memoization and loading states.
